chore(governance): Phase D Port SSOT and Env Hygiene Seal
