+++ .env.example
@@ -30,3 +30,10 @@ ENABLE_ACTION_RAIL_APPROVAL=false
 # AWS_REGION=eu-central-1
 # AWS_BUCKET_NAME=santis-media-global
 # CDN_DOMAIN=https://cdn.santis.club
+
+# --- Port SSOT Registry (Phase D) ---
+# Admin Panel: 8080
+# Ingestion API: 3030
+# Marketing/Static Site: 8081
+# Telemetry Signaling: 4040
+# Fastify Internal API: 3000
