+++ admin-panel/.env.example
@@ -1,3 +1,3 @@
 VITE_TELEMETRY_WS_URL=ws://localhost:4040
 VITE_TELEMETRY_API_URL=http://localhost:4040
-VITE_CORE_API_URL=http://localhost:8080
+VITE_CORE_API_URL=http://localhost:3030
