+++ apps/ingestion-api/.env.example
@@ -1,0 +1,9 @@
+# Ingestion API (Sovereign Bus Backend)
+PORT=3030
+NODE_ENV=development
+
+# Database
+# DATABASE_URL=postgresql://user:pass@localhost:5432/santis
+
+# Security
+ADMIN_SECRET_TOKEN=replace_with_min_24_char_token
