+++ apps/ingestion-api/src/config/websocket-gateway.config.ts
@@ -1,11 +1,11 @@
 import { z } from "zod";
 
 const WebSocketGatewayEnvSchema = z.object({
-  WS_PORT: z.coerce.number().int().min(1024).max(65535).default(8080),
+  WS_PORT: z.coerce.number().int().min(1024).max(65535).default(4040),
   WS_HOST: z.string().min(1).default("0.0.0.0"),
   WS_PATH: z.string().startsWith("/").default("/ws"),
   WS_ALLOWED_ORIGINS: z.string()
-    .default("http://localhost:5173,http://127.0.0.1:5173,http://localhost:5500,http://127.0.0.1:5500,http://localhost:3030,http://127.0.0.1:3030")
+    .default("http://localhost:8080,http://127.0.0.1:8080,http://localhost:8081,http://127.0.0.1:8081,http://localhost:3030,http://127.0.0.1:3030")
     .transform(str => str.split(",").map(s => s.trim()).filter(Boolean)),
   WS_ALLOWED_ORIGIN_PATTERNS: z.string()
     .optional()
